Move rng to separate file
