+++ Skybox.cpp
@@ -8,6 +8,51 @@ using namespace Entities;
 using Renderer::VertexArray;
 using Renderer::CubeMap;
 
+const std::vector<f32> SKYBOX_VERTICES =
+{
+	-SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
+	-SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
+	SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
+	SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
+	SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
+	-SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
+
+	-SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
+	-SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
+	-SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
+	-SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
+	-SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
+	-SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
+
+	SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
+	SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
+	SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
+	SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
+	SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
+	SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
+
+	-SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
+	-SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
+	SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
+	SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
+	SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
+	-SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
+
+	-SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
+	SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
+	SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
+	SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
+	-SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
+	-SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
+
+	-SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
+	-SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
+	SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
+	SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
+	-SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
+	SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE
+};
+
 constexpr std::array<const char*, 6> SKYBOX_TEXTURE_FILES =
 {
 	"gfx/right.png",
@@ -19,52 +64,7 @@ constexpr std::array<const char*, 6> SKYBOX_TEXTURE_FILES =
 };
 
 Skybox::Skybox()
-	: cubeMap(std::make_shared<CubeMap>(SKYBOX_TEXTURE_FILES))
+	: vao(std::make_shared<VertexArray>(3, SKYBOX_VERTICES)),
+	  cubeMap(std::make_shared<CubeMap>(SKYBOX_TEXTURE_FILES))
 {
-	const std::vector<f32> SKYBOX_VERTICES =
-	{
-		-SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
-		-SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
-		SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
-		SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
-		SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
-		-SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
-
-		-SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
-		-SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
-		-SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
-		-SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
-		-SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
-		-SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
-
-		SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
-		SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
-		SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
-		SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
-		SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
-		SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
-
-		-SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
-		-SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
-		SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
-		SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
-		SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
-		-SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
-
-		-SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
-		SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
-		SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
-		SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
-		-SKYBOX_SIZE, SKYBOX_SIZE, SKYBOX_SIZE,
-		-SKYBOX_SIZE, SKYBOX_SIZE, -SKYBOX_SIZE,
-
-		-SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
-		-SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
-		SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
-		SKYBOX_SIZE, -SKYBOX_SIZE, -SKYBOX_SIZE,
-		-SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE,
-		SKYBOX_SIZE, -SKYBOX_SIZE, SKYBOX_SIZE
-	};
-
-	vao = std::make_shared<VertexArray>(3, SKYBOX_VERTICES);
 }
