+++ Util.h
@@ -4,8 +4,6 @@
 // Utility header file
 // Contains utility macros and defines
 
-#include <random>
-
 #include <cstdint>
 #include <cstddef>
 #include <cassert>
@@ -35,28 +33,5 @@ using usize = std::size_t;
 
 // Handy global variables
 inline f32 g_Delta = 1.0f;
-
-namespace Util
-{
-	// Random number between a range
-	template<typename T>
-	T Rand_Range(T min, T max)
-	{
-		static thread_local std::mt19937_64 generator(777);
-		std::uniform_real_distribution<T> distributer(min, max);
-		return distributer(generator);
-	}
-
-	// Truly random number between a range
-	template<typename T>
-	T True_Rand_Range(T min, T max)
-	{
-		std::random_device rd;
-		std::seed_seq ss{ rd(), rd() };
-		static thread_local std::mt19937_64 generator(ss);
-		std::uniform_real_distribution<T> distributer(min, max);
-		return distributer(generator);
-	}
-}
 
 #endif
